refactor: rename file to clarify intent
