+++ packages/nextclade/src/__tests__/CMakeLists.txt
@@ -9,7 +9,7 @@ find_package(Microsoft.GSL 3.1.0 REQUIRED)
 find_package(fmt 7.1.0 REQUIRED)
 
 add_executable(${PROJECT_NAME}
-  analyzeSequence.test.cpp
+  findNucChanges.test.cpp
   findNucleotideRanges.test.cpp
   findPrivateMutations.test.cpp
   getAminoacidChanges.test.cpp
