+++ packages/nextclade/src/__tests__/findNucChanges.test.cpp
@@ -1,3 +1,5 @@
+#include "../analyze/findNucChanges.h"
+
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
 
@@ -6,7 +8,6 @@
 
 #include "../../include/nextclade/nextclade.h"
 #include "../../include/nextclade/private/nextclade_private.h"
-#include "../analyze/analyze.h"
 
 #define EXPECT_ARR_EQ(expected, actual) ASSERT_THAT(actual, ::testing::ElementsAreArray(expected));
