+++ packages/nextclade/src/analyze/findNucChanges.cpp
@@ -1,4 +1,4 @@
-#include "analyze.h"
+#include "findNucChanges.h"
 
 #include <vector>
